% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{addToBase}
\alias{addToBase}
\title{Extend the base formula with variables we are adjusting for.}
\usage{
addToBase(base_formula, adjustingVariables)
}
\arguments{
\item{base_formula}{a formula object with the dependent variable as a function of the independent variable}

\item{adjustingVariables}{a character vector of column names from the data.frame to adjust for}
}
\value{
An updated formula object representing a base formula with adjustment variables.
}
\description{
Extend the base formula with variables we are adjusting for.
}

