% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{analyze_linear_mod}
\alias{analyze_linear_mod}
\title{analyze_linear_mod}
\usage{
analyze_linear_mod(formula, dat, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{dat}{the data.frame to perform analysis on.}
}
\value{
A data.frame representing the linear association study.
}
\description{
Main worker function to perform linear associations.
}
\examples{
\dontrun{
analyze_linear_mod()
}

}

