% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{analyze_logistic_mod}
\alias{analyze_logistic_mod}
\title{analyze_logistic_mod}
\usage{
analyze_logistic_mod(formula, dat, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{dat}{the data.frame to perform analysis on.}
}
\value{
A data.frame representing the binary outcome study.
}
\description{
Main worker function to perform binary outcome associations.
}
\examples{
\dontrun{
analyze_logistic_mod()
}

}

