% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{analyze_surival_mod}
\alias{analyze_surival_mod}
\title{analyze_survival_mod}
\usage{
analyze_surival_mod(formula, dat, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{dat}{the data.frame to perform analysis on.}
}
\value{
A data.frame representing the survival outcome study.
}
\description{
Main worker function to perform survival analysis. This functionality is still in BETA TESTING.
}
\examples{
\dontrun{
analyze_survival_mod()
}

}

