% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{is.survey}
\alias{is.survey}
\title{Check if the object is a survey design object from the 'survey' library.}
\usage{
is.survey(design)
}
\arguments{
\item{design}{is an object to test for}
}
\value{
A boolean if the design argument provided is a 'survey.design' object or not.
}
\description{
Check if the object is a survey design object from the 'survey' library.
}

