% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{is_variable_categorical}
\alias{is_variable_categorical}
\title{Determines if column in the main data.frame is categorical.}
\usage{
is_variable_categorical(mainTab, varname)
}
\arguments{
\item{mainTab}{is a data.frame representing the data to be analyzed.}

\item{varname}{is the name of a column of the data.frame to be determined if it is a categorical variable or not based on the number of factors.}
}
\value{
A factorization of the column of the data.frame.
}
\description{
Determines if column in the main data.frame is categorical.
}
\examples{

}

