% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{logistic_mod}
\alias{logistic_mod}
\title{Main worker function to perform binary outcome associations.}
\usage{
logistic_mod(formula, dat, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{dat}{the data.frame to perform analysis on.}
}
\value{
A data.frame representing the binary outcome study.
}
\description{
Main worker function to perform binary outcome associations.
}
\examples{
\dontrun{
logistic_mod()
}
}

