% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{manhattan}
\alias{manhattan}
\title{manhattan}
\usage{
manhattan(data, qthreshold = 0.05, nlabels = 6, ...)
}
\arguments{
\item{data}{is the vector containing a list of p-values to display in a manhattan plot.}

\item{qthreshold}{is the cutoff of significance (the q-value).}

\item{nlabels}{is the number of top hits to label.}
}
\value{
A plot object.
}
\description{
Takes a vector of p-values and then displays in a manhattan plot with additional options. Usually can take the output of an xwas result
without further processing in most cases.
}
\examples{
\dontrun{
manhattan()
manhattan(glucose, main="XWAS of LBXGLU")
}

}

