% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{NHANES survey results.}
\format{A data frame with 10122 rows and 939 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.nhanes.org/}
}
\usage{
data(nhanes)
}
\description{
A dataset containing the NHANES survey information.
}
\examples{
data(nhanes)
}
\keyword{datasets}

