% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{qqplot}
\alias{qqplot}
\title{qqplot}
\usage{
qqplot(data, xlab = expression(-log[10] * "(p-expected)"),
  ylab = expression(-log[10] * "(p-empirical)"), pch = 20, ...)
}
\arguments{
\item{data}{is an object either a numeric character vector of p-values or the output of the xwas function from which that information can be derived.}

\item{xlab}{is a default label for the x-axis passed on to the plot function.}

\item{ylab}{is a default label for the y-axis passed on to the plot function.}

\item{pch}{is set to 20 by default which is a small bullet point passed on to the plot function.}

\item{...}{are additional arguments passed on to the plot function.}
}
\value{
none
}
\description{
Takes a data object that is either the output of the
}
\examples{
\dontrun{
qqplot()
qqplot(data)
}

}

