% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{unskew}
\alias{unskew}
\title{http://www.r-statistics.com/2013/05/log-transformations-for-skewed-and-wide-distributions-from-practical-data-science-with-r/}
\usage{
unskew(x)
}
\description{
http://www.r-statistics.com/2013/05/log-transformations-for-skewed-and-wide-distributions-from-practical-data-science-with-r/
}

