% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(data, qthreshold = 0.05, xlab = "Effect",
  ylab = expression(-log[10] * "(p-value)"), pch = 20, ...)
}
\arguments{
\item{data}{is the vector containing a list of p-values to display in a manhattan plot.}

\item{qthreshold}{is the cutoff of significance (the q-value).}

\item{xlab}{is a default x-axis label passed on to the plot function.}

\item{ylab}{is a default y-axis label passed on to the plot function.}

\item{pch}{is default to 20 which is a small bullet passed onto the plot function.}

\item{...}{are additional arguments passed onto the plot function.}
}
\value{
none
}
\description{
Takes a list of linear regressions and displays a volcano plot.
}
\examples{
\dontrun{
volcano()
volcano(data)
}

}

