% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{xlm}
\alias{xlm}
\title{xlm}
\usage{
xlm(data, depvar = NULL, varname = NULL, adjvars = c(), permute = 0,
  categorical = 0, verbose = TRUE)
}
\arguments{
\item{data}{is a data.frame containing the data to perform analysis on.}

\item{depvar}{a character vector with a column name for the dependent variable.}

\item{varname}{a character vector with a column name for the independent variable.}

\item{adjvars}{a list of variables to adjust for in the regression.}

\item{permute}{is an optional parameter of how many permutations to perform in the bootstrap.}

\item{categorical}{is a binary option representing whether or not the variable is categorical.}

\item{verbose}{is a boolean that determines if we print extra information. Not recommend for an XWAS, only one off analyses.}
}
\value{
A data.frame object representing the regression.
}
\description{
Performs a singular linear regression or binary outcomes association study. It requires you to specify the exact column variable 
and other permutation details. It does a lot of blind analysis and requires the xwas function to determine other meta-characteristics.
}
\examples{
\dontrun{
xlm()
xlm(data=nhanes, depvar="LBXGLU", varname="LBXGTC", adjvars=c("female", "RIDAGEYR"), permute=10)
}

}

