% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{xwas.to.df}
\alias{xwas.to.df}
\title{xwas.to.df}
\usage{
xwas.to.df(data, ...)
}
\arguments{
\item{data}{is a list of data.frame objects that represent the regression tests, usually this is the output from the xwas function.}

\item{...}{are additional parameters that are passed on to the p.adjust function.}
}
\value{
a data.frame with the p-value, estimate effect size, and adjusted p-value.
}
\description{
Takes the output of the xwas function that returns a list of regressions and parses the results into a simple data.frame. This is useful for downstream plotting within the xwas library as well as custom implementations.
}
\examples{
\dontrun{
xwas.to.df(data)
}

}

