% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{linear_mod}
\alias{linear_mod}
\title{linear_mod}
\usage{
linear_mod(formula, dat, design = NULL, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{dat}{the data.frame to perform analysis on, can be optional if the design argument is used.}

\item{design}{a survey::svydesign object that has the experimental design.}
}
\value{
A data.frame representing the linear association study.
}
\description{
Main worker function to perform linear associations.
}
\examples{
\dontrun{
linear_mod()
}

}

