% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{surival_mod}
\alias{surival_mod}
\title{survival_mod}
\usage{
surival_mod(formula, data, design = NULL, ...)
}
\arguments{
\item{formula}{an R formula class object.}

\item{data}{the data.frame to perform analysis on, can be option if design argument is used.}

\item{design}{a survey::svydesign object that has the experimental design.}

\item{...}{are additional arguments passed on to the regression.}
}
\value{
A data.frame representing the survival outcome study.
}
\description{
Main worker function to perform survival analysis.
}
\examples{
\dontrun{
survival_mod()
}

}

