% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{xwas}
\alias{xwas}
\title{Performs X-Wide Association Analysis.}
\usage{
xwas(data, depvar, varname, adjvars = c(), permute = 0, categorical = 0)
}
\arguments{
\item{data}{is a data.frame containing the data to perform analysis on.}

\item{depvar}{a character vector with a column name for the dependent variable.}

\item{varname}{a character vector with a column name for the independent variable.}

\item{adjvars}{a list of variables to adjust for in the regression.}

\item{permute}{is an optional parameter of how many permutations to perform in the bootstrap.}

\item{categorical}{is a binary option representing whether or not the variable is categorical.}
}
\value{
A data.frame object representing the regression.
}
\description{
Performs X-Wide Association Analysis.
}
\examples{
\dontrun{
xwas()
}

}

