% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{xwas}
\alias{xwas}
\title{xwas}
\usage{
xwas(data, depvar, timevar = NULL, varname = NULL, adjvars = c(),
  design = NULL, permute = 0, n = 1, verbose = TRUE)
}
\arguments{
\item{data}{is the data.frame containing what to analyze.}

\item{depvar}{is required and the outcome or dependent variable we are looking to analyze in the context of multiple factors.}

\item{timevar}{is optional and if provided will be used in conjunction with the depvar to create a survival::Survey object for survival analysis.}

\item{varname}{is optional and the independent variable, without being specified we will consider all variables. If it is provided there will only be 1 regression test performed.}

\item{adjvars}{is optional and is a vector of a set of variables to adjust for under every condition, if not specified we will scan all variables without adjusting on a first pass.}

\item{design}{a survey::svydesign object that has the experimental design.}

\item{permute}{is the number of times to bootstrap each variable analysis in the xwas.}

\item{n}{is the number of cores to use for the multi-core implementation, value must be > 1 or set to "MAX".}

\item{verbose}{is a boolean to print extra information.}
}
\value{
A data.frame representing the output.
}
\description{
Used to systematically perform an xlm across all variables. This function contains a lot of the logic for a systematic analysis
including logic to determine variable characteristics.
}
\examples{
\dontrun{
xwas()
xwas(data=nhanes, depvar="LBXGLU")
xwas(data=nhanes, depvar="LBXGLU", adjvars=c("BMXBMI", "RIDAGEYR", "female"))
}

}

